+++ Droids/src/main/Vector2.java
@@ -4,27 +4,27 @@ public class Vector2 {
 
 	/* FIELD */
 
-	private float x, y;
+	private double x, y;
 
 	/* GET SET */
 
-	public float getX() {
+	public double getX() {
 		return x;
 	}
 
-	public void setX(float x) {
+	public void setX(double x) {
 		this.x = x;
 	}
 
-	public float getY() {
+	public double getY() {
 		return y;
 	}
 
-	public void setY(float y) {
+	public void setY(double y) {
 		this.y = y;
 	}
 
-	public void setLocation(float x, float y) {
+	public void setLocation(double x, double y) {
 		this.x = x;
 		this.y = y;
 	}
@@ -35,7 +35,7 @@ public class Vector2 {
 		this(0f, 0f);
 	}
 
-	public Vector2(float x, float y) {
+	public Vector2(double x, double y) {
 		this.x = x;
 		this.y = y;
 	}
